requirements.txt
opencv-python~=4.5.2.52
numpy~=1.24.3
torch~=1.13.1+cu117
tensorflow-gpu~=2.8.0
pillow~=9.5.0
matplotlib~=3.7.1
pyod~=0.9.9
torchvision~=0.14.1+cu117
scikit-learn~=1.2.2
tqdm~=4.65.0
cleanlab~=2.0.0
pandas~=2.0.1
xlwt~=1.3.0
torchtext~=0.14.1
scipy~=1.10.1
keras~=2.8.0
wavemix~=0.2.4
einops~=0.6.1
pyyaml~=6.0
deepod~=0.2.0